# StringCenterNode/docs/api-spec.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "StringCenter",
    "description": "Tab sharing social media platform"
  },
  "host": "cs-309-ss-5.cs.iastate.edu:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/get-user": {
      "get": {
        "summary": "Get all users",
        "description": "Returns a list containing all users.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "A list of users",
            "schema": {
              "properties": {
                "users": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "No users found",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/get-user/{username}": {
      "get": {
        "summary": "Get user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The username for fetching user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Username not found",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Create a new user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Username, password, and confirmPassword are required",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "confirmPassword": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid request. The response key \"field\" will be the type of error, for example \"username\" if the error occured in the username. The defualt key is \"message\".",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tab" : {
      "get": {
        "summary": "Get all tabs",
        "description": "Returns a list containing all tabs.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "A list of tabs",
            "schema": {
              "properties": {
                "tabs": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Tab"
                  }
                }
              }
            }
          },
          "400": {
            "description": "No tabs found",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tab/createTab": {
      "post": {
        "summary": "Create a new user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Tab object",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Tab"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Tab"
            }
          },
          "400": {
            "description": "Invalid tab. The response key \"field\" will be the type of error, the default key is \"message\".",
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string"
        }
      }
    },
    "Note": {
      "type": "string"
    },
    "InstString": {
      "type": "object",
      "properties": {
        "notes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Note"
          }
        },
        "tuning": {
          "type": "string"
        }
      }
    },
    "Measure": {
      "type": "object",
      "required": [
        "info",
        "tuning",
        "stringCount",
        "strings"
      ],
      "properties": {
        "info": {
          "type": "string"
        },
        "tuning": {
          "type": "string"
        },
        "stringCount": {
          "type": "integer"
        },
        "strings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InstString"
          }
        }
      }
    },
    "Tab": {
      "type": "object",
      "required": [
        "author_username",
        "info",
        "measureCount",
        "measures"
      ],
      "properties": {
        "author_username": {
          "type": "string"
        },
        "info": {
          "type": "string"
        },
        "measureCount": {
          "type": "integer"
        },
        "measures": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Measure"
          }
        }
      }
    }
  }
}